+++ src/helpers.js
@@ -7,7 +7,7 @@ const generateUUID = function f_GenerateUUID () {
         d = Math.floor(d/16);
         return (c=='x' ? r : (r&0x3|0x8)).toString(16);
     });
-    return uuid;
+    return 'pv_' + uuid;
 }
 
 const helpers = {